refactoring of nlu tools
