+++ modules/nlu/src/backend/tools/math.ts
@@ -1,6 +1,14 @@
 import _ from 'lodash'
 
+/**
+ * Vectorial distance between two N-dimentional points
+ * a[] and b[] must be of same dimention
+ */
 export function ndistance(a: number[], b: number[]): number {
+  if (a.length !== b.length) {
+    throw new Error(`Can't calculate distance between vectors of different length (${a.length} vs ${b.length})`)
+  }
+
   let total = 0
   for (let i = 0; i < a.length; i++) {
     const diff = b[i] - a[i]
@@ -9,50 +17,7 @@ export function ndistance(a: number[], b: number[]): number {
   return Math.sqrt(total)
 }
 
-/**
- * Returns the levenshtein distance between two strings
- * @returns the # of operations required to go from a to b
- */
-export function levenshtein(a: string, b: string): number {
-  let tmp
-
-  if (a.length === 0) {
-    return b.length
-  }
-  if (b.length === 0) {
-    return a.length
-  }
-  if (a.length > b.length) {
-    tmp = a
-    a = b
-    b = tmp
-  }
-
-  let i: number,
-    j: number,
-    res: number = 0
-
-  const alen = a.length,
-    blen = b.length,
-    row = Array(alen)
-
-  for (i = 0; i <= alen; i++) {
-    row[i] = i
-  }
-
-  for (i = 1; i <= blen; i++) {
-    res = i
-    for (j = 1; j <= alen; j++) {
-      tmp = row[j - 1]
-      row[j - 1] = res
-      res = b[i - 1] === a[j - 1] ? tmp : Math.min(tmp + 1, Math.min(res + 1, row[j] + 1))
-    }
-  }
-
-  return res
-}
-
-export function GetZPercent(z) {
+export function GetZPercent(z: number) {
   if (z < -6.5) {
     return 0.0
   }
